# Closes #3033: Optimize CSV write